Extract BundleMetrics into DetectiveCore and split CLIOutput
- New DetectiveCore.BundleMetrics with size(at:) and format(bytes:). Replaces three near-identical implementations: CLI's bundleSize/format, MetadataLoaderService's getTotalBundleSize/calculateDirectorySize/formatSize, and a now-removed local copy.
- Split AppDetectiveCLI.swift: arg parsing and main() stay in AppDetectiveCLI.swift; the Output Encodable struct and the text/json/error printers move to a new CLIOutput.swift (renamed CLIOutput, namespaced under CLIPrinter).
