+++ AppDetective/AppDetective/Services/MetadataLoaderService.swift
@@ -1,4 +1,5 @@
 import Combine
+import DetectiveCore
 import SwiftUI
 
 class MetadataLoaderService {
@@ -71,9 +72,8 @@ class MetadataLoaderService {
         autoreleasepool {
             let iconData = NSWorkspace.shared.icon(forFile: path)
                 .tiffRepresentation
-            let sizeInBytes = getTotalBundleSize(atPath: path)
-            let sizeString =
-                sizeInBytes != nil ? formatSize(bytes: sizeInBytes!) : nil
+            let sizeInBytes = BundleMetrics.size(at: URL(fileURLWithPath: path))
+            let sizeString = sizeInBytes.map(BundleMetrics.format(bytes:))
 
             var thumbnailData: Data?
             if let fullData = iconData, let fullImage = NSImage(data: fullData) {
@@ -107,55 +107,4 @@ class MetadataLoaderService {
         return thumbnail.tiffRepresentation
     }
 
-    private func getTotalBundleSize(atPath path: String) -> Int64? {
-        let url = URL(fileURLWithPath: path).resolvingSymlinksInPath()
-        do {
-            let resourceKeys: Set<URLResourceKey> = [
-                .totalFileSizeKey, .totalFileAllocatedSizeKey,
-            ]
-            let resourceValues = try url.resourceValues(forKeys: resourceKeys)
-
-            if let totalSize = resourceValues.totalFileSize {
-                return Int64(totalSize)
-            } else if let allocatedSize = resourceValues.totalFileAllocatedSize {
-                return Int64(allocatedSize)
-            } else {
-                return calculateDirectorySize(atPath: path)
-            }
-        } catch {
-            return calculateDirectorySize(atPath: path)
-        }
-    }
-
-    private func calculateDirectorySize(atPath path: String) -> Int64? {
-        let url = URL(fileURLWithPath: path).resolvingSymlinksInPath()
-        guard let enumerator = FileManager.default.enumerator(
-            at: url, includingPropertiesForKeys: [.fileSizeKey],
-            options: [],
-            errorHandler: { _, error -> Bool in
-                print("[MetadataLoader] Enumerator error: \(error)")
-                return true
-            })
-        else {
-            return nil
-        }
-
-        var totalSize: Int64 = 0
-        for case let fileURL as URL in enumerator {
-            do {
-                let resourceValues = try fileURL.resourceValues(forKeys: [.fileSizeKey])
-                totalSize += Int64(resourceValues.fileSize ?? 0)
-            } catch {
-                print("[MetadataLoader] Error getting size for \(fileURL.path): \(error)")
-            }
-        }
-        return totalSize
-    }
-
-    private func formatSize(bytes: Int64) -> String {
-        let formatter = ByteCountFormatter()
-        formatter.allowedUnits = [.useKB, .useMB, .useGB]
-        formatter.countStyle = .file
-        return formatter.string(fromByteCount: bytes)
-    }
 }
